refactor: seperate MyApplication type
